Convert to ES6
* Replace requireJS with Webpack + BabelJS
