+++ basic/src/main/resources/static/app.js
@@ -1,0 +1,70 @@
+'use strict';
+
+// tag::vars[]
+const React = require('react');
+const client = require('./client');
+// end::vars[]
+
+// tag::app[]
+class App extends React.Component {
+
+	constructor(props) {
+		super(props);
+		this.state = {employees: []};
+	}
+
+	componentDidMount() {
+		client({method: 'GET', path: '/api/employees'}).done(response => {
+			this.setState({employees: response.entity._embedded.employees});
+		});
+	}
+
+	render() {
+		return (
+			<EmployeeList employees={this.state.employees}/>
+		)
+	}
+}
+// end::app[]
+
+// tag::employee-list[]
+class EmployeeList extends React.Component{
+	render() {
+		var employees = this.props.employees.map(employee =>
+			<Employee key={employee._links.self.href} employee={employee}/>
+		);
+		return (
+			<table>
+				<tr>
+					<th>First Name</th>
+					<th>Last Name</th>
+					<th>Description</th>
+				</tr>
+				{employees}
+			</table>
+		)
+	}
+}
+// end::employee-list[]
+
+// tag::employee[]
+class Employee extends React.Component{
+	render() {
+		return (
+			<tr>
+				<td>{this.props.employee.firstName}</td>
+				<td>{this.props.employee.lastName}</td>
+				<td>{this.props.employee.description}</td>
+			</tr>
+		)
+	}
+}
+// end::employee[]
+
+// tag::render[]
+React.render(
+	<App />,
+	document.getElementById('react')
+)
+// end::render[]
+
