+++ basic/src/main/resources/static/client.js
@@ -1,22 +1,19 @@
-define(function(require) {
-	'use strict';
+'use strict';
 
-	var rest = require('rest');
-	var defaultRequest = require('rest/interceptor/defaultRequest');
-	var mime = require('rest/interceptor/mime');
-	var uriTemplateInterceptor = require('./api/uriTemplateInterceptor');
-	var errorCode = require('rest/interceptor/errorCode');
-	var baseRegistry = require('rest/mime/registry');
+var rest = require('rest');
+var defaultRequest = require('rest/interceptor/defaultRequest');
+var mime = require('rest/interceptor/mime');
+var uriTemplateInterceptor = require('./api/uriTemplateInterceptor');
+var errorCode = require('rest/interceptor/errorCode');
+var baseRegistry = require('rest/mime/registry');
 
-	var registry = baseRegistry.child();
+var registry = baseRegistry.child();
 
-	registry.register('text/uri-list', require('./api/uriListConverter'));
-	registry.register('application/hal+json', require('rest/mime/type/application/hal'));
+registry.register('text/uri-list', require('./api/uriListConverter'));
+registry.register('application/hal+json', require('rest/mime/type/application/hal'));
 
-	return rest
+module.exports = rest
 		.wrap(mime, { registry: registry })
 		.wrap(uriTemplateInterceptor)
 		.wrap(errorCode)
 		.wrap(defaultRequest, { headers: { 'Accept': 'application/hal+json' }});
-
-});
