+++ events/pom.xml
@@ -22,9 +22,6 @@
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<java.version>1.8</java.version>
-		<spring-data-releasetrain.version>Gosling-RELEASE</spring-data-releasetrain.version>
-		<spring-hateoas.version>0.19.0.RELEASE</spring-hateoas.version>
-		<spring.version>4.2.1.RELEASE</spring.version>
 	</properties>
 
 	<dependencies>
@@ -39,6 +36,10 @@
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-data-rest</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-devtools</artifactId>
 		</dependency>
 		<!-- tag::websocket[] -->
 		<dependency>
@@ -108,7 +109,13 @@
                             <arguments>install</arguments>
                         </configuration>
                     </execution>
-                </executions>
+					<execution>
+						<id>webpack build</id>
+						<goals>
+							<goal>webpack</goal>
+						</goals>
+					</execution>
+				</executions>
             </plugin>
 
 			<plugin>
@@ -122,27 +129,5 @@
 
 		</plugins>
 	</build>
-
-	<repositories>
-		<repository>
-			<id>spring-milestones</id>
-			<name>Spring Milestones</name>
-			<url>http://repo.spring.io/milestone</url>
-			<snapshots>
-				<enabled>false</enabled>
-			</snapshots>
-		</repository>
-	</repositories>
-
-	<pluginRepositories>
-		<pluginRepository>
-			<id>spring-milestones</id>
-			<name>Spring Milestones</name>
-			<url>http://repo.spring.io/milestone</url>
-			<snapshots>
-				<enabled>false</enabled>
-			</snapshots>
-		</pluginRepository>
-	</pluginRepositories>
 
 </project>
