+++ events/src/main/resources/static/package.json
@@ -4,7 +4,7 @@
   "description": "Demo of ReactJS + Spring Data REST",
   "repository": {
     "type": "git",
-    "url": "git@github.com:gregturn/reactjs-and-spring-data-rest.git"
+    "url": "git@github.com:spring-guides/tut-reactjs-and-spring-data-rest.git"
   },
   "keywords": [
     "rest",
@@ -16,13 +16,20 @@
   "author": "Greg L. Turnquist",
   "license": "ASLv2",
   "bugs": {
-    "url": "https://github.com/gregturn/reactjs-and-spring-data-rest/issues"
+    "url": "https://github.com/spring-guides/tut-reactjs-and-spring-data-rest/issues"
   },
-  "homepage": "https://github.com/gregturn/reactjs-and-spring-data-rest",
+  "homepage": "https://github.com/spring-guides/tut-reactjs-and-spring-data-rest",
   "dependencies": {
-    "react-tools": "^0.13.3"
+    "babel-core": "^5.8.25",
+    "babel-loader": "^5.3.2",
+    "react-tools": "^0.13.3",
+    "webpack": "^1.12.2"
   },
   "devDependencies": {
-    "bower": "^1.3.12"
+    "bower": "^1.3.12",
+    "bower-webpack-plugin": "^0.1.9"
+  },
+  "scripts": {
+    "watch": "webpack --watch -d"
   }
 }
