+++ events/src/main/resources/static/websocket-listener.js
@@ -1,21 +1,19 @@
-define(function(require) {
-	'use strict';
+'use strict';
 
-	var SockJS = require('sockjs-client'); // <1>
-	require('stomp-websocket'); // <2>
+var SockJS = require('sockjs-client'); // <1>
+require('stomp-websocket'); // <2>
 
-	return {
-		register: register
-	};
-
-	function register(registrations) {
-		var socket = SockJS('/payroll'); // <3>
-		var stompClient = Stomp.over(socket);
-		stompClient.connect({}, function(frame) {
-			registrations.forEach(function (registration) { // <4>
-				stompClient.subscribe(registration.route, registration.callback);
-			});
+function register(registrations) {
+	var socket = SockJS('/payroll'); // <3>
+	var stompClient = Stomp.over(socket);
+	stompClient.connect({}, function(frame) {
+		registrations.forEach(function (registration) { // <4>
+			stompClient.subscribe(registration.route, registration.callback);
 		});
-	}
+	});
+}
 
-});
+module.exports = {
+	register: register
+};
+
