+++ hypermedia/src/main/resources/static/bower.json
@@ -1,7 +1,7 @@
 {
   "name": "spring-data-rest-and-reactjs",
   "version": "0.1.0",
-  "homepage": "https://github.com/gregturn/reactjs-and-spring-data-rest",
+  "homepage": "https://github.com/spring-guides/tut-reactjs-and-spring-data-rest",
   "authors": [
     "Greg Turnquist <gturnquist@pivotal.io>"
   ],
@@ -27,8 +27,9 @@
     "tests"
   ],
   "dependencies": {
-    "requirejs": "~2.1.20",
+    "react": "~0.13.3",
     "rest": "~1.3.1",
-    "requirejs-react-jsx": "~0.14.2"
+    "sockjs-client": "~1.0.3",
+    "stomp-websocket": "~2.3.4"
   }
 }
