+++ hypermedia/src/main/resources/templates/index.html
@@ -3,12 +3,13 @@
 <head lang="en">
     <meta charset="UTF-8"/>
     <title>ReactJS + Spring Data REST</title>
-    <script async="" data-main="/run.js" src="/bower_components/requirejs/require.js"></script>
     <link rel="stylesheet" href="/main.css" />
 </head>
 <body>
 
     <div id="react"></div>
+
+    <script src="built/bundle.js"></script>
 
 </body>
 </html>
