--- security/src/main/resources/static/main.js
@@ -1,10 +1,0 @@
-define(function(require) {
-	'use strict';
-
-	document.addEventListener('DOMContentLoaded', new function() {
-
-		require('jsx!app');
-
-	}, false);
-
-});
